feat: wip of passkey user registration
