+++ apps/operator/package.json
@@ -12,14 +12,16 @@
     "@jsonforms/react": "^3.2.1",
     "@jsonforms/vanilla-renderers": "^3.2.1",
     "@nextui-org/react": "^2.2.10",
-    "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-popover": "^1.0.7",
-    "@repo/dally": "workspace:*",
-    "@repo/ui": "workspace:*",
-    "@repo/codegen": "workspace:*",
-    "@types/jsonwebtoken": "^9.0.6",
     "@novu/node": "^0.24.1",
     "@novu/notification-center": "^0.24.1",
+    "@radix-ui/react-accordion": "^1.1.2",
+    "@radix-ui/react-popover": "^1.0.7",
+    "@repo/codegen": "workspace:*",
+    "@repo/dally": "workspace:*",
+    "@repo/ui": "workspace:*",
+    "@simplewebauthn/browser": "^10.0.0",
+    "@simplewebauthn/server": "^10.0.0",
+    "@types/jsonwebtoken": "^9.0.6",
     "graphql-request": "latest",
     "graphql-tag": "^2.12.6",
     "jwt-decode": "^4.0.0",
