+++ apps/operator/src/app/api/auth/[...nextauth]/route.ts
@@ -2,5 +2,5 @@
  * Export the route handlers from our auth config
  * and set the runtime to the Edge
  */
-export { GET, POST } from '@/lib/auth'
+export { GET, POST } from '@/lib/auth/auth'
 export const runtime = 'edge'
