+++ apps/operator/src/app/api/auth/registration-verification/route.ts
@@ -1,0 +1,25 @@
+import { NextResponse } from 'next/server'
+
+export async function POST(request: Request) {
+  const bodyData = await request.json()
+
+  const fData = await fetch(
+    `${process.env.API_REST_URL}/v1/registration/verification`,
+    {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify(bodyData),
+      credentials: 'include',
+    },
+  )
+
+  if (fData.ok) {
+    return NextResponse.json(await fData.json(), { status: 200 })
+  }
+
+  if (fData.status !== 201) {
+    return NextResponse.json(await fData.json(), { status: fData.status })
+  }
+}
