+++ apps/operator/src/app/api/auth/verify/route.ts
@@ -1,10 +1,11 @@
 import { type NextRequest, NextResponse } from 'next/server'
 
-
 export async function GET(request: NextRequest) {
   const searchParams = request.nextUrl.searchParams
   const token = searchParams.get('token')
-  const fData = await fetch(`${process.env.API_REST_URL}/v1/verify?token=${token}`)
+  const fData = await fetch(
+    `${process.env.API_REST_URL}/v1/verify?token=${token}`,
+  )
 
   if (fData.ok) {
     return NextResponse.json(await fData.json(), { status: 200 })
