+++ apps/operator/src/middleware.ts
@@ -4,7 +4,7 @@
  */
 
 import { NextResponse } from 'next/server'
-import { auth } from './lib/auth'
+import { auth } from './lib/auth/auth'
 
 export default auth((req) => {
   /**
